requirements.txt
backports.functools-lru-cache==1.4
cycler==0.10.0
decorator==4.1.2
distro==1.0.4
dlib==19.7.0
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
sqlite3
annoy
flask
